+++ extensions/facebook/src/main/java/app/morphe/extension/facebook/settings/FamilyNames.java
@@ -23,6 +23,7 @@ public final class FamilyNames {
     public static final String SPONSORED_STORIES = "Hide sponsored stories";
     public static final String STORY_AUTO_ADVANCE = "Stop Story auto-advance";
     public static final String SPONSORED_REELS = "Hide sponsored reels";
+    public static final String REEL_DECLUTTER = "Clean up Reels";
     public static final String EXTERNAL_BROWSER = "Open links in external browser";
     public static final String SANITIZE_SHARING_LINKS = "Sanitize sharing links";
     public static final String STORY_DOWNLOAD = "Download any story";